requirements.txt
opencv_python==4.4.0.46
numpy==1.19.4
pytesseract==0.3.6
scikit_image==0.17.2
torchvision==0.8.1+cu101
xlwt==1.3.0
matplotlib==3.3.2
torch==1.7.0+cu101
python_bidi==0.4.2
six==1.14.0
pathlib2==2.3.5
Pillow==8.0.1
PyYAML==5.3.1
skimage==0.0
